5 Functions has been move  in serateted fiile Functions.tsx
